# Show the elapsed time of single test and speedup some tests