Following #10038.

This PR introduces two changes.
1. Show the elapsed time of a single test in the test output, in order to have a more detailed understanding of the changes in test run time.

    Output:
    ```
    [ok]: PFADD without arguments creates an HLL value (1 ms)
    [ok]: Approximated cardinality after creation is zero (0 ms)
    [ok]: PFADD returns 1 when at least 1 reg was modified (0 ms)
    [ok]: PFADD returns 0 when no reg was modified (0 ms)
    ```

2. Speedup two tests related to `key-load-delay` configuration.
    By comparing the output of the following thress CIs, other tests do not seem to be affected by #10003.

| Test    | Before #10003  | After #10003   | After this PR |
| ---------- | ------------------------ | ------------------------ | -------- |
| client freed during loading | 9 seconds    | 22 seconds | 7 seconds |
| config during loading | 4 seconds | 15 seconds  | 4 seconds |

### Compare CI:
before #10003: https://github.com/sundb/redis/runs/4712829066?check_suite_focus=true
after #10003: https://github.com/sundb/redis/runs/4679660840?check_suite_focus=true
after this PR: https://github.com/sundb/redis/runs/4712795765?check_suite_focus=true